-Removing size_t from sound functions

+++ Examples/StereoKitCTest/demo_mic.cpp
@@ -31,7 +31,7 @@ size_t sample_count    = 0;
 void demo_mic_init() {
 	mic_input = mic_start();
 
-	for (size_t i = 0; i < mic_device_count(); i++) {
+	for (int32_t i = 0; i < mic_device_count(); i++) {
 		mic_device_names.push_back(mic_device_name(i));
 	}
 
@@ -90,7 +90,7 @@ void demo_mic_update() {
 
 		// And draw the volume indicator!
 		mic_intensity = mic_intensity + (mic_intensity_dest - mic_intensity) * time_elapsedf() * 16;
-		float    scale = 0.1f + 0.1*mic_intensity;
+		float    scale = 0.1f + 0.1f*mic_intensity;
 		color128 color = { 1,1,1, fmaxf(0.1f,mic_intensity) };
 		render_add_mesh(mic_visual_mesh, mic_visual_mat, matrix_trs({0,0,-0.5f}, quat_identity, vec3_one*scale), color);
 		sprite_draw    (mic_sprite, matrix_trs({ -0.03f,0.03f,-0.5f }, quat_identity, vec3_one * 0.06f));
